extra

+++ console/migrations/m180215_161627_create_Extrafields_tables.php
@@ -16,43 +16,43 @@ class m180215_161627_create_Extrafields_tables extends Migration
             'id'    => $this->primaryKey(),
             'name'  => $this->string()->notNull(),
             'slug'  => $this->string()->defaultValue(null),
-            'type'  => $this->number(11),            
+            'type'  => $this->integer(11),            
         ]);
 
         $this->createTable('extrafield_field_set', [
             'id'        => $this->primaryKey(),
-            'set_id'    => $this->number(11)->notNull(),
-            'field_id'  => $this->number(11)->notNull(),
-            'is_requred' => $this->number(3)->notNull(),            
+            'set_id'    => $this->integer(11)->notNull(),
+            'field_id'  => $this->integer(11)->notNull(),
+            'is_requred' => $this->integer(3)->notNull(),            
         ]);
 
         $this->createTable('extrafield_input', [
             'id'            => $this->primaryKey(),
             'object_type'   => $this->string(255)->defaultValue(null),
             'value'         => $this->string(255)->defaultValue(null),
-            'object_id'     => $this->number(11)->notNull(),  
-            'field_id'      => $this->number(11)->notNull(),
+            'object_id'     => $this->integer(11)->notNull(),  
+            'field_id'      => $this->integer(11)->notNull(),
         ]);
 
         $this->createTable('extrafield_input_list', [
             'id'        => $this->primaryKey(),
-            'field_id'  => $this->number(11)->defaultValue(null),
-            'value'     => $this->number(11)->defaultValue(null),
-            'object_id' => $this->number(11)->defaultValue(null),
+            'field_id'  => $this->integer(11)->defaultValue(null),
+            'value'     => $this->integer(11)->defaultValue(null),
+            'object_id' => $this->integer(11)->defaultValue(null),
             'object_type' => $this->string(255)->defaultValue(null),
         ]);
 
         $this->createTable('extrafield_input_list_defined', [
             'id'        => $this->primaryKey(),
-            'field_id'  => $this->number(11)->defaultValue(null),
+            'field_id'  => $this->integer(11)->defaultValue(null),
             'value'     => $this->string(255)->defaultValue(null),
         ]);
 
         $this->createTable('extrafield_int', [
             'id'        => $this->primaryKey(),
-            'field_id'  => $this->number(11)->defaultValue(null),
-            'value'     => $this->number(11)->defaultValue(null),
-            'object_id' => $this->number(11)->defaultValue(null),
+            'field_id'  => $this->integer(11)->defaultValue(null),
+            'value'     => $this->integer(11)->defaultValue(null),
+            'object_id' => $this->integer(11)->defaultValue(null),
             'object_type' => $this->string(255)->defaultValue(null),
         ]);
 
@@ -64,8 +64,8 @@ class m180215_161627_create_Extrafields_tables extends Migration
 
         $this->createTable('extrafield_text', [
             'id'        => $this->primaryKey(),
-            'field_id'  => $this->number(11)->defaultValue(null),
-            'object_id' => $this->number(11)->defaultValue(null),
+            'field_id'  => $this->integer(11)->defaultValue(null),
+            'object_id' => $this->integer(11)->defaultValue(null),
             'value'     => $this->text(),
             'object_type' => $this->string(255)->defaultValue(null),
         ]);
